Add files via upload
Problem: This query is basically to find a murderer and the villain who sponsored the killing;
A  crime has taken place and the detective needs my help as a DA to do his work with findings from the gathered result files. The detective handed over the crime scene report, but I somehow lost it. Vaguely I remember that the crime was a ​murder​ that occurred sometime on ​Jan.15, 2018,​ and that it took place in ​SQL City​. To help out I wrote this query to find the murderer and the villain behind the killing.